refactor: moved some logic from Grid to ./utils/grid.js
